+++ src/settings/MappingConfiguration/utils/index.js
@@ -1,2 +1,3 @@
+export * from './getFieldNameOptions';
 export * from './getTranslatorOptions';
 export * from './mappings';