+++ src/settings/components/FieldTranslateDefault/FieldTranslateDefault.js
@@ -1,6 +1,5 @@
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-
 import { Field } from 'react-final-form';
 
 import { Checkbox } from '@folio/stripes/components';