+++ src/settings/components/index.js
@@ -1,5 +1,6 @@
 export { FieldDefaultValue } from './FieldDefaultValue';
 export { FieldFromFieldName } from './FieldFromFieldName';
+export { FieldFromPath } from './FieldFromPath';
 export { FieldTranslateDefault } from './FieldTranslateDefault';
 export { FieldTranslator } from './FieldTranslator';
 export { MappingFieldCard } from './MappingFieldCard';