+++ CHANGELOG.md
@@ -5,4 +5,6 @@
 * Module setup. Refs UIGS-1.
 * Add GOBI integration area to FOLIO settings. Refs UIGS-2.
 * Display GOBI integration Configuration files and mappings. Refs UIGS-4.
+* Allow user to edit mappings. Refs UIGS-5.
 * Allow user to reset to default configuration. Refs UIGS-6.
+* Define From field name and Path and allow Custom path option. Refs UIGS-8.